blk-crypto: introduce blk_crypto_bio_sectors_alignment()
Add a function which returns the number of sectors to which the size of
the given bio (and any bios split from it) must be aligned based on its
encryption context.  For a bio with an encryption context, this will be
the data unit size of the encryption context in sectors.

This is needed to allow the block layer to ensure that the lengths of
split bios remain aligned to the crypto data unit size, not merely to
the disk's logical block size which may be smaller.

Signed-off-by: Satya Tangirala <satyat@google.com>
[EB: improved comments and commit message]
Signed-off-by: Eric Biggers <ebiggers@google.com>