block: introduce bio_required_sector_alignment()
Add a function which returns the number of sectors to which the size of
the given bio (and any bios split from it) must be aligned.  This will
be logical_block_size for bios without an encryption context, or the
greater of logical_block_size and the crypto data_unit_size for bios
with an encryption context.

This is needed to allow the block layer to ensure that the lengths of
split bios remain aligned to the crypto data unit size, not merely to
the disk's logical block size which may be smaller.

Signed-off-by: Satya Tangirala <satyat@google.com>
[EB: improved comment and commit message]
Signed-off-by: Eric Biggers <ebiggers@google.com>