block: warn if split bio won't be aligned to crypto data unit size
The number of sectors passed to bio_split() must be aligned to
bio_required_sector_alignment() in order to ensure that bios with an
encryption context aren't split in the middle of a crypto data unit.
All callers other than bounce.c have been updated to ensure this, so add
a WARN_ON_ONCE() if the number of sectors is not properly aligned.

(bounce.c was not updated because it is legacy code.  Any device that
enables bounce buffering won't declare inline encryption support, so
bounce.c should never see a bio with an encryption context.)

Signed-off-by: Satya Tangirala <satyat@google.com>
Signed-off-by: Eric Biggers <ebiggers@google.com>