blk-crypto-fallback: handle data unit split across multiple bvecs
In order to support direct I/O on encrypted files as well as possibly
other future use cases such as filesystem metadata encryption, we need
to relax the restriction that if a bio has an encryption context, then
all its bvecs must be be aligned to the encryption context's data unit
size.  Instead we'll just require that the bio's total size be aligned
to the data unit size (in addition to the usual logical_block_size
alignment requirement on bvecs).

Update blk-crypto-fallback accordingly to no longer assume that bvecs
are data_unit_size aligned.  This requires two changes:

- When splitting an overly-long bio, round the size of the of the first
  part of the bio down to the nearest data unit boundary rather than
  assuming that a bvec boundary is sufficient.

- When encrypting/decrypting a bio's data, instead of iterating through
  each data unit in each bvec (which is no longer valid), rework the
  code to iterate through the data units and advance through the bvecs
  as needed.  Note that a data unit may now be split among multiple
  bvecs, and in this case we need to create a multi-element scatterlist
  to pass to the crypto API.

Co-developed-by: Satya Tangirala <satyat@google.com>
Signed-off-by: Satya Tangirala <satyat@google.com>
Signed-off-by: Eric Biggers <ebiggers@google.com>