lib/crypto: sha512: Deduplicate the round constants
Almost all the SHA-512 implementations in lib/crypto/ need the 640-byte
round constants array, and the couple that don't need it have a fallback
to the generic code which needs it anyway.  So it is always needed.
However, usually multiple copies of it are being included in the
resulting binary, e.g. 2 copies on both arm64 and x86_64.  (x86_64 would
have 4 copies, but it uses mergeable sections to get it down to 2.)

Since the lib/crypto/ SHA-512 code is now a single module, it's now
fairly straightforward to define the round constants array once in
sha512.c and reference that same copy from assembly.  So, let's do that.

Signed-off-by: Eric Biggers <ebiggers@kernel.org>