lib/crc: x86/crc32c: Specialize the VPCLMULQDQ+AVX512 code for CRC-32C
Modify crc-pclmul-template.S and crc32-pclmul.S to generate a function
crc32c_vpclmul_avx512() which is specialized for the needs of crc32c().
Compared to crc32_lsb_vpclmul_avx512() which is designed to handle any
least-significant-bit-first 32-bit CRC, it specializes the final
reduction for CRC-32C by using crc32q instructions instead of a generic
method.  It also omits the code that handles len < 256, since this is
not needed for crc32c() which already uses different code for len < 512.

Then, make crc32c() use crc32c_vpclmul_avx512() instead of
crc32_lsb_vpclmul_avx512().

The result is slightly improved crc32c() performance when this function
is used: len >= 512 on CPUs with a good implementation of AVX-512.

Signed-off-by: Eric Biggers <ebiggers@kernel.org>