crypto: hctr2 - Drop support for asynchronous xctr ciphers
There are no asynchronous xctr ciphers, so there is no need to support
them.  Simplify and streamline the code in the synchronous case, i.e.
the case that actually matters, instead.

For example, by no longer having to support resuming the calculation
after an asynchronous return of the xctr cipher, we can just keep more
of the state on the stack instead of in the request context.

Signed-off-by: Eric Biggers <ebiggers@kernel.org>